# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jelajah Lokal API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/public/articles": {
      "get": {
        "summary": "List Articles",
        "operationId": "list_articles_api_public_articles_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "default": 9,
              "title": "Pagesize"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Featured"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArticleList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/articles/{slug}": {
      "get": {
        "summary": "Article Detail",
        "operationId": "article_detail_api_public_articles__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_PublicArticleDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/articles": {
      "get": {
        "summary": "List Articles",
        "operationId": "list_articles_api_admin_articles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_list_ArticleDraft__"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Article",
        "operationId": "create_article_api_admin_articles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Data_ArticleDraft_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/editorial/generate": {
      "post": {
        "summary": "Editorial Generate",
        "operationId": "editorial_generate_api_admin_editorial_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/editorial/apply": {
      "post": {
        "summary": "Editorial Apply",
        "operationId": "editorial_apply_api_admin_editorial_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplyInput": {
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          },
          "selectedFields": {
            "items": {
              "type": "string",
              "enum": [
                "excerpt",
                "metaDescription",
                "tags"
              ]
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Selectedfields"
          },
          "values": {
            "$ref": "#/components/schemas/ApplyValues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "selectedFields",
          "values"
        ],
        "title": "ApplyInput"
      },
      "ApplyResponse": {
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          },
          "articleDocumentId": {
            "type": "string",
            "format": "uuid",
            "title": "Articledocumentid"
          },
          "applied": {
            "type": "boolean",
            "const": true,
            "title": "Applied"
          },
          "draftRevision": {
            "type": "integer",
            "title": "Draftrevision"
          },
          "selectedFields": {
            "items": {
              "type": "string",
              "enum": [
                "excerpt",
                "metaDescription",
                "tags"
              ]
            },
            "type": "array",
            "title": "Selectedfields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "articleDocumentId",
          "applied",
          "draftRevision",
          "selectedFields"
        ],
        "title": "ApplyResponse"
      },
...
      "ArticleCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "title": "Title",
            "default": ""
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "title": "Slug",
            "default": ""
          },
          "body": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Paragraph"
                },
                {
                  "$ref": "#/components/schemas/Heading"
                },
                {
                  "$ref": "#/components/schemas/Quote"
                },
                {
                  "$ref": "#/components/schemas/ListBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "heading": "#/components/schemas/Heading",
                  "image": "#/components/schemas/ImageBlock",
                  "list": "#/components/schemas/ListBlock",
                  "paragraph": "#/components/schemas/Paragraph",
                  "quote": "#/components/schemas/Quote"
                }
              }
            },
            "type": "array",
            "maxItems": 300,
            "title": "Body"
          },
          "excerpt": {
            "type": "string",
            "maxLength": 240,
            "title": "Excerpt",
            "default": ""
          },
          "metaDescription": {
            "type": "string",
            "maxLength": 160,
            "title": "Metadescription",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "categoryDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categorydocumentid"
          },
          "authorDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authordocumentid"
          },
          "coverMediaDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Covermediadocumentid"
          },
          "coverAlt": {
            "type": "string",
            "maxLength": 180,
            "title": "Coveralt",
            "default": ""
          },
          "coverCaption": {
            "type": "string",
            "maxLength": 240,
            "title": "Covercaption",
            "default": ""
          },
          "imageCredit": {
            "type": "string",
            "maxLength": 240,
            "title": "Imagecredit",
            "default": ""
          },
          "imageSourceUrl": {
            "type": "string",
            "maxLength": 2048,
            "title": "Imagesourceurl",
            "default": ""
          },
          "sourceLinks": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Sourcelinks"
          },
          "informationCheckedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Informationcheckedat"
          },
          "operationalClaims": {
            "type": "boolean",
            "title": "Operationalclaims",
            "default": false
          },
          "regionLabel": {
            "type": "string",
            "maxLength": 80,
            "title": "Regionlabel",
            "default": ""
          },
          "featured": {
            "type": "boolean",
            "title": "Featured",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ArticleCreate"
      },
      "ArticleDraft": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "title": "Title",
            "default": ""
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "title": "Slug",
            "default": ""
          },
          "body": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Paragraph"
                },
                {
                  "$ref": "#/components/schemas/Heading"
                },
                {
                  "$ref": "#/components/schemas/Quote"
                },
                {
                  "$ref": "#/components/schemas/ListBlock"
                },
                {
                  "$ref": "#/components/schemas/ImageBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "heading": "#/components/schemas/Heading",
                  "image": "#/components/schemas/ImageBlock",
                  "list": "#/components/schemas/ListBlock",
                  "paragraph": "#/components/schemas/Paragraph",
                  "quote": "#/components/schemas/Quote"
                }
              }
            },
            "type": "array",
            "maxItems": 300,
            "title": "Body"
          },
          "excerpt": {
            "type": "string",
            "maxLength": 240,
            "title": "Excerpt",
            "default": ""
          },
          "metaDescription": {
            "type": "string",
            "maxLength": 160,
            "title": "Metadescription",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "categoryDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categorydocumentid"
          },
          "authorDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authordocumentid"
          },
          "coverMediaDocumentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Covermediadocumentid"
          },
          "coverAlt": {
            "type": "string",
            "maxLength": 180,
            "title": "Coveralt",
            "default": ""
          },
          "coverCaption": {
            "type": "string",
            "maxLength": 240,
            "title": "Covercaption",
            "default": ""
          },
          "imageCredit": {
            "type": "string",
            "maxLength": 240,
            "title": "Imagecredit",
            "default": ""
          },
          "imageSourceUrl": {
            "type": "string",
            "maxLength": 2048,
            "title": "Imagesourceurl",
            "default": ""
          },
          "sourceLinks": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Sourcelinks"
          },
          "informationCheckedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Informationcheckedat"
          },
          "operationalClaims": {
            "type": "boolean",
            "title": "Operationalclaims",
            "default": false
          },
          "regionLabel": {
            "type": "string",
            "maxLength": 80,
            "title": "Regionlabel",
            "default": ""
          },
          "featured": {
            "type": "boolean",
            "title": "Featured",
            "default": false
          },
          "documentId": {
            "type": "string",
            "format": "uuid",
            "title": "Documentid"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedat"
          },
          "publishedRevision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedrevision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documentId",
          "revision",
          "createdAt",
          "updatedAt",
          "publishedAt",
          "publishedRevision"
        ],
        "title": "ArticleDraft"
      },
...
      "Data_ArticleDraft_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ArticleDraft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[ArticleDraft]"
      },
...
      "Data_PublicArticleDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicArticleDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[PublicArticleDetail]"
      },
...
      "Data_list_ArticleDraft__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ArticleDraft"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Data[list[ArticleDraft]]"
      },
...
      "GenerateInput": {
        "properties": {
          "articleDocumentId": {
            "type": "string",
            "format": "uuid",
            "title": "Articledocumentid"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid",
            "title": "Idempotencykey"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "articleDocumentId",
          "idempotencyKey"
        ],
        "title": "GenerateInput"
      },
      "GenerateResponse": {
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid",
            "title": "Requestid"
          },
          "articleDocumentId": {
            "type": "string",
            "format": "uuid",
            "title": "Articledocumentid"
          },
          "revisionFingerprint": {
            "type": "string",
            "title": "Revisionfingerprint"
          },
          "result": {
            "$ref": "#/components/schemas/GeneratedMetadata"
          },
          "cacheHit": {
            "type": "boolean",
            "title": "Cachehit"
          },
          "quota": {
            "$ref": "#/components/schemas/QuotaDTO"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "articleDocumentId",
          "revisionFingerprint",
          "result",
          "cacheHit",
          "quota"
        ],
        "title": "GenerateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Heading": {
        "properties": {
          "type": {
            "type": "string",
            "const": "heading",
            "title": "Type"
          },
          "level": {
            "type": "integer",
            "enum": [
              2,
              3
            ],
            "title": "Level"
          },
          "children": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextInline"
                },
                {
                  "$ref": "#/components/schemas/LinkInline"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "link": "#/components/schemas/LinkInline",
                  "text": "#/components/schemas/TextInline"
                }
              }
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Children"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "level",
          "children"
        ],
        "title": "Heading"
      },
      "ImageBlock": {
        "properties": {
          "type": {
            "type": "string",
            "const": "image",
            "title": "Type"
          },
          "mediaDocumentId": {
            "type": "string",
            "format": "uuid",
            "title": "Mediadocumentid"
          },
          "alt": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Alt"
          },
          "caption": {
            "type": "string",
            "maxLength": 240,
            "title": "Caption",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "mediaDocumentId",
          "alt"
        ],
        "title": "ImageBlock"
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "pageCount": {
            "type": "integer",
            "title": "Pagecount"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "pageCount",
          "total"
        ],
        "title": "PageMeta"
      },
      "Paragraph": {
        "properties": {
          "type": {
            "type": "string",
            "const": "paragraph",
            "title": "Type"
          },
          "children": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextInline"
                },
                {
                  "$ref": "#/components/schemas/LinkInline"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "link": "#/components/schemas/LinkInline",
                  "text": "#/components/schemas/TextInline"
                }
              }
            },
            "type": "array",
            "maxItems": 200,
            "title": "Children"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "children"
        ],
        "title": "Paragraph"
      },
...
      "PublicArticleCard": {
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid",
            "title": "Documentid"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "category": {
            "$ref": "#/components/schemas/PublicName"
          },
          "author": {
            "$ref": "#/components/schemas/PublicName"
          },
          "cover": {
            "$ref": "#/components/schemas/Cover"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Publishedat"
          },
          "readingTimeMinutes": {
            "type": "integer",
            "title": "Readingtimeminutes"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documentId",
          "slug",
          "title",
          "excerpt",
          "category",
          "author",
          "cover",
          "publishedAt",
          "readingTimeMinutes",
          "featured"
        ],
        "title": "PublicArticleCard"
      },
      "PublicArticleDetail": {
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid",
            "title": "Documentid"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "category": {
            "$ref": "#/components/schemas/PublicName"
          },
          "author": {
            "$ref": "#/components/schemas/PublicName"
          },
          "cover": {
            "$ref": "#/components/schemas/Cover"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Publishedat"
          },
          "readingTimeMinutes": {
            "type": "integer",
            "title": "Readingtimeminutes"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured"
          },
          "body": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Paragraph"
                },
                {
                  "$ref": "#/components/schemas/Heading"
                },
                {
                  "$ref": "#/components/schemas/Quote"
                },
                {
                  "$ref": "#/components/schemas/ListBlock"
                },
                {
                  "$ref": "#/components/schemas/PublicImageBlock"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "heading": "#/components/schemas/Heading",
                  "image": "#/components/schemas/PublicImageBlock",
                  "list": "#/components/schemas/ListBlock",
                  "paragraph": "#/components/schemas/Paragraph",
                  "quote": "#/components/schemas/Quote"
                }
              }
            },
            "type": "array",
            "title": "Body"
          },
          "metaDescription": {
            "type": "string",
            "title": "Metadescription"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "sourceLinks": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "type": "array",
            "title": "Sourcelinks"
          },
          "informationCheckedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Informationcheckedat"
          },
          "coverCaption": {
            "type": "string",
            "title": "Covercaption"
          },
          "imageCredit": {
            "type": "string",
            "title": "Imagecredit"
          },
          "imageSourceUrl": {
            "type": "string",
            "title": "Imagesourceurl"
          },
          "regionLabel": {
            "type": "string",
            "title": "Regionlabel"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documentId",
          "slug",
          "title",
          "excerpt",
          "category",
          "author",
          "cover",
          "publishedAt",
          "readingTimeMinutes",
          "featured",
          "body",
          "metaDescription",
          "tags",
          "sourceLinks",
          "informationCheckedAt",
          "coverCaption",
          "imageCredit",
          "imageSourceUrl",
          "regionLabel",
          "updatedAt"
        ],
        "title": "PublicArticleDetail"
      },
      "PublicArticleList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PublicArticleCard"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PublicArticleList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}